[FIX] web_studio: prevent field creation in new button
When creating a new button (for the buttonbox), one has to select a field
(many2one pointing to the current model). We want to prevent the field
creation at this step to simplify the process. The same applies when selecting
the paperformat in report.

The previous hack `nodeOptions.no_create` correctly hides the `Create` link
below the text but after the focus out, the create modal still appears.