[FIX] web_studio: update sidebar after edition
When editing a node (ex: field label), the operation is applied on the view followed
by a view re-rendering. As the node has been modified, the sidebar needs to be updated.
To achieve this, we save the selected node and click on it again.

Since rev https://github.com/odoo-dev/enterprise/commit/3aabad806efe30a9d0c1a7961140e2d7b665cc00
this selected node is reset in the mixin and set back on the corresponding click handler.

The mixin handler thus needs to be called before the other handler.