[FIX] hr_appraisal,website_version: read_group
The models HrAppraisal and Experiment override read_group to
handle the case when it is grouped by field state, which is of
type selection. However, they set as value for each group the
array [key, name] of the selection option, which is incorrect
(it should set key as value, like the default read_group does,
and as the name is known client side anyway).

This case was handled by the old views, but it isn't by the new
ones, which are less permissive. We choose to fix the read_group
override instead of allowing both cases in JS, as it is always
better to have a clear and single API.