[IMP] sale_subscription: no more delegation on analytic accounts
For historical reasons, the sale.subscription model used the delegation
inheritance mechanism on analtyic accounts. This limited the use of the
analytic accounting since changing the analytic account of a
subscription after the fact was problematic (loss of history if any
invoice line was already created). Furthermore, installing the module
enabled the analytic accounting for all users, which is weird.

Removing the delegation inheritance also has the added advnatage of
splitting the access rights/rules of these models.