[IMP] helpdesk: Allow to merge helpdesk tickets
Purpose
=======

Implement the merge feature  for the tickets, same as we did for the tasks.

same mechanism (log in chatter, merge of the description, archiving the deprecated tickets)

Specification
=============

The purpose of this wizard is, allow user to merge the tickets into existing ticket or new ticket
once tickets get merged, following action will be performed.
   - message will automatically posted on merged tickets as well as on target ticket
   - merged tickets will be archived
   - followers of the merged tickets will be added in target ticket
   - description of the merged tickets will be added in target ticket
   - tags of the merged tickets will be added in target ticket