[IMP] sale_subscription(_asset): move ugly '_fields' introspection
This sale.order override used to be in website_subscription before
that particular feature was moved down to sale_subscription. The
rationale at the time was not to create a link module for overriding
a single value in a dict; however this does not apply anymore since
the sale_contract_asset module can do the job properly.

While I'm at it, let's add some multi-company enforcement here.