[ADD] web_studio: allow adding rainbow effect to buttons
In this commit, we add a feature to web_studio: we can add an 'effect'
to buttons in form view.  Editing this value means that the web client
will display a 'rainbow' effect when the button is clicked and the
action is completed.

joint work with dja, adr, and ged