[ADD] helpdesk_timesheet: log timesheet on ticket
This commit provide 2 new modules to handle timesheets
on helpdesk tickets :

- helpdesk_timesheet : by linking a project to a team,
an employee will be able to log timesheet directly on
the ticket. However, the ticket can be link to a task
to be sell later if needed.
The billing process is handled by the timesheet on task
flow (sale_timesheet).
Additionnal config are avaiblable on helpdesk team.

- helpdesk_sale_timesheet: even if the billing process for
ticket timesheets is in sale_timesheet, this bridge module
display a warning on the ticket when the delivered quantites
is overdued.