[FIX] web_studio: clear caches on changes
The ModelFieldSelector keeps an internal cache for fields_get.
This cache needs to be cleared in studio, alongside the one of
the data_manager, after editions.

A new 'clear_cache' event has been introduced to clear both the
ModelFieldSelector and data_manager caches. We use this new
mechanism in studio to ensure that the ModelFieldSelector cache
is also cleared on changes.