[FIX] helpdesk: team kanban display
In the kanban view of helpdesk.team, both links

"0 Tickets to Review"
"0 Unassigned Tickets"

need to be on one line, instead of having a linebreak after the number.

It used to be that way but a change in the styling framework changed
the css display option.

We use <t t-esc=...> instead of a <field name=...> to avoid creating the
extra span that caused the linebreak.