[FIX] web_studio: report_editor: display message on error
Some reports require additional information to render. When
accessed in Odoo, the user has to fill a wizard to provide those
information. However, when these reports are opened in Studio, we
don't pass through the configuration wizard. So before this rev.,
'Internal server error' was displayed instead of the report
content.

Combined with a community commit odoo/odoo@0d2fb541e, this rev. detects
that the report hasn't been rendered because of missing information,
and it displays a proper error message instead.

Note that those reports requiring additional information are no
longer used in enterprise (they are only used in community), so
it's probably not a real use case to edit them with Studio.

opw-746000