[FIX] web_grid: don't erase domain when adding a line
When a line is added, the grid view is reloaded. Before this rev.,
the created record was passed as argument to the reload function,
which considered it as the params. Since the record contained a
'domain' attribute (empty array), the reload function erased the
current domain with the given one, and thus the current domain was
lost.