[FIX] account_reports: xlsx sheet name too long
An error was raised when exporting to xlsx with some translations.
This error was due to worksheet names too long (more than 31 chars).

A patch was made in community, at odoo/odoo@c1e154be, along with this one, to
override the add_worksheet method from xlsxwriter.Workbook. The patched method
now removes unwanted characters and troncate names that are too long.

The regular xlsxwriter is still imported in case the community version
is not found. That way, if someone does not update community, it will
not raise an error.

Closes #1169
opw-751157