[FIX] website_sign: get location coordinates
JSON.stringify now doesn't work on google chrome Coordinates object
(previously it already didn't work on IE and firefox) because the values
are now on the object prototype (as the standard recommends).

Also, the code doesn't expect keys like "altitude" and so on, so we
have to cherry-pick only those we wish for.

opw-746710
closes #1179