[IMP] web_studio: Normalize custom views (#1178)
When adding or removing custom elements to a view in studio, odoo adds
xpaths on top of the existing Studio arch, doesn't matter if it's
removing an already existing element, removing an element added through
studio, adding or modifying elements.

This causes the resulting arch to become overcrowded with useless statements.

i.e.: Adding field x after field y, and then deleting field x, all
within web_studio, will create 2 xpaths, one for creating and one for
deleting.

The expected behaviour would be an empty arch since we deleted the
element we created.

Solution:

WebStudioController now comes with a normalize method that will be
executed while editing a view in Studio, this method will go through the
whole studio arch and append new elements to existing xpaths if it can,
or delete them if there's an xpath that would replace an existing one.

This makes for a more compact and logical studio view, and also works on
legacy customized views if edited. The method will fall back to the old
way of creating one xpath per operation if the normalized view would
crash.