[IMP] web_clearbit: only suggest for companies
Before this revision, the clearbit widget was activated for the
res.partner model, irrespective of whether we are dealing with a company
or an individual. In terms of usability, it does not make sense to be
prompted with company names when we are filling in the name of an
individual.

The following commit brings the following changes:

- Activate the suggestion on input ony if we are dealing with a company;
- As the previous change introduces res.partner specific logic, we
  initialize web_clearbit only if the model is res.partner.
- The associated QUnit test is adapted to the previous two requirements
  and now tests the behaviour of clearbit for a company and an
  individual.