[FIX] website_sign: multi-line item wrap issue
When editing a multi-line text item on a document, a textarea is used so
we have preformatted text and the text is wrapping if the line is longer
than the available width.

But when previewing or printing the pdf this is not the case. Only
explicit new line are kept and line are not wrapped automatically.

This would be a bigger effort to implement the feature (a try has been
made https://gist.github.com/nle-odoo/1bea6ad1eecd738c39f969808f467011
but this currently has its own shortfalls) so this pull request just
modify the edition so it reflects what appear when previewing and
printing.

With this commit, a line return must be inserted imperatively in the
edition to have content on several line.

opw-746900
closes #1191