[FIX] web_studio: Use static view in tests
The tests for the XML normalization of views modified in Studio were
using res.partner_view_form as its view for unit tests, changes to the
view done either in the community or enterprise version of Odoo could
potentially break the tests (and it did at some point).

This is fixed by using a static version (hardcoded string) of a
modified/minified res.partner_view_form view, so that even if the view
changes elsewhere, the tests won't be broken.