[FIX] intercompany_rules: Wrong conversion in multi-currency inter-company invoices
Steps to reproduce:
-You need to be in muti-currency multi-company setup.
-One company is in EUR, the other one is in USD.
-In Intercompany settings, set "Create invoice/refund" for both companies.
-Make sure you have active exchange rates for EUR and USD and that one x-rate
is different from 1 (for example, USD x-rate = 1.2, EUR x-rate = 1).
-Make sure both companies' partners are defined as customer and vendor.

-In the US company, create a customer invoice for the EUR company,
with amount = 1000 USD.
-It will create a vendor bill in the EUR company, which is good. But the
vendor bill will show an amount of 1200 USD, which doesn't make sense.

Fix: When an intercompany invoice is created no conversion of the price unit
must be done. Before the fix it was done in function "_onchange_product_id"
defined in model "account.invoice.line" in community.

opw:751186