[FIX] mrp_account : cost report server error
Cost report was defined in old way that was no
more standart and thus the view couln't be found.

The report should be declared with a name format:
report.model_name.report_name. Also the view could not be found
because the module that contains them was not indicated.

This commit also update the report template with the new standart
python dict instead of the old json declaration.