[FIX] delivery_fedex: allow rating of very lightweight items
Some users may want to ship very lightweight items; in order to give
them a rating, we round the weight of the shipping to the smallest
value accepted by FedEx: 0.01 kg or lb.

In the case where the weight is exactly 0.0 because weights are not set
on products, ignore this in order to be able to return an error message
"Missing package weight".