[FIX] (sale_)account_taxcloud: Set `Cart Item Index` from 1 instead of line.id
Purpose
=======

Uncertain how to "get" to the number of items, but in talking with taxcloud, an "item index" is sent over with the api call which states how many cart items are there.  In the attached example, you can see that this item index is showing as 42,594.  They do not accept integers over 32,767 so the validation fails.  I suppose to recreate you would set the cart index greater than 32,767 and then attempt to validate your order.

Specification
=============

The cart index should be reset on each new sales order or invoice, so if you have 4 product lines, the index would be 4.