[IMP] web_enterprise: Burger Menu for mobile
This commit improves the UX of the menu (navbar) in mobile.

Basically, there is a Burger icon in the top right of the screen,
and no more User Menu nor Company Switcher in the navbar. When the
burger icon is clicked, a menu opens fullscreen. It contains the
submenus of the current app if any. The user menu and company
switcher are also available from this new menu.