[FIX] voip: not working on standard user accounts
When a user, like most normal users, does not have access to the
Settings menu, it has no access to the ir config parameters of
the voip, so he will not be able to dial (in production mode).

That is why we add a sudo in the function that gets these parameters.
That function returns those values to the web client however, so we need
to check  that the user is not a portal user.  (has the group
base.group_user)

opw-751585