[FIX] web_studio: open a grouped kanban
It was not possible to open Studio with a grouped kanban since the new views.
When grouped, the kanban state `data` is a `list` datapoint (and not `record`).
This was not taken into account when selecting the first record.

The previous code also modified the `data` in place while this object can be
shared. Here, the object is not modified and the rendering function only take
the first record into account.

Some tests have been added to cover this use case.