[FIX] mrp_account : cost report server error on product template
This commit fixes the product_template_cost_structure report
that has been forgetten in the following commit
601885cfa0d71982d18493997d7874e4b1983394.

This commit forces the use of the report_name field on the
ir.actions.report as the qweb view to be used when rendering a report,
so we clone the original cost_structure view. We also fix the model name
so that it can be found by the report engine.