[FIX] account_reports: Allow a Billing user to open the Customer Statements
PURPOSE
=======

When clicking on the menuitem customer statements, the followup that haven't been created yet are created to be displayed. As the billing user doesn't have the right to create an account_report_manager record, this commit gives this right to the Billing user to allows him to perform this action.