[FIX] account_reports: self.env.user.company_id.currency_id.rate
1. Install the "Accounting" app.
2. Enter the "General Settings" and enable "Multi-company and
   Multi-currency" options.
3. Create 2 Companies (or modify the current one + add a new one) with
   different currencies (in my case, I used USD for the parent company &
   EGP for the child company).
4. Create rates for the used currencies (in my case as of 1/1/2017: USD
   for the parent company is 1.00 and EGP is 18.00, USD for the child
   company is 0.055556 and EGP is 1.00).
5. Create a user: Demo with Accountant access rights and select his
   company as the parent company and give him Allowed companies as both
   parent and child companies.
   Modify the administrator to have allowed companies too to the parent
   and child companies.
6. Create Customer Invoice with the parent company in USD, and a
   Customer Invoice with the child company in EGP (after populating the
   default chart of accounts template for both companies).
7. Look onto the P & L report from the perspective of the "Demo" (with
   the parent company selected as main company) user while changing the
   main company of the administrator, the values will differ based on
   the currency of the main company of the admin!!

The core of the issue is that we are calling
`self.env.user.company_id.currency_id.rate`.

In method `_compute_current_rate`, if no company is provided in the
context, we retrieve the company of the user thanks to
`self.env.user.company_id`. However, since `user` is browsed as `sudo`,
the `self.env.user` is the admin in the mentioned method. Indeed, we are
doing:
`self.env.user.company_id.currency_id.env.user.company_id.rate`

A proper fix would be to avoid browsing the user as `sudo`, but that
cannot be done in stable branch.

opw-753139