[IMP] project_forecast: allow forecast for day off
Now, working calendar of employee is required. So, we
don't need to take this case into account in computed
field method.
Also, when planning a forecast on a day off/holidays
period for the employee, we don't want to raise an
exception, but set the forecast to zero percentage
available.