[FIX] web_gantt: exclude field is broken
On gantt view, a field can be set as 'consolidation_exclude',
but it does not work because the js code iterate on the name
of the field. Setting 'exclude' field on gatt view for forecast
will create warning in server log, as it try to read each letter
of field name as field to read (forecast.read(['e','x','c',...]).