[FIX] account_reports: protect __builtins__ against early overriding
Before this commit, when computing a financial report, FormulaContext might be asked for another item than the flavours of sum
which accidentally overrode some __builtins__ functions. (in this opw's case, it was abs() that was overriden, messing things up in the safe_eval later on ).
after this commit, we ensure that the key asked to FormulaContext doesn't belong to the builtins either

opw 756708

Closes #1244