[FIX] website_sign: make on-the-fly email detection better
When choosing the person to whom a signature request will be sent, the
user has the possibility to enter the name and email of that person in
an unique input and the system will understand that a res.partner with
that name and email has to be created. This worked well, unless the
email contained multiple dots (".") after the "@" (as in test@a.b.com).

This commit improves the regex used for email detection to allow such
email addresses and also simplifies this regex.

opw-765739