[MOV] (website_)helpdesk: move customer portal to helpdesk
As for account and sale modules it is now possible to display tickets
in the customer portal without relying on website module. This commit
contains notably :

 * move access_token field helpdesk.ticket model as it is used to grant
   access to tickets to customers;
 * move code related to notification groups and access action as token
   based feature is now accessible in helpdesk;
 * move portal access rights and rules;
 * update customer portal templates to use portal features and not website
   features anymore;

website helpdesk still contains published mixin for teams as well as
templates used to display teams on website, with action button to
create tickets, and/or have access to forums and slides. Those features
are still website features and therefore stay in website_helpdesk.