[FIX] web: don't save empty required x2manys
Before this commit, the same function was called at the rendering and at the writing of a field.
It caused a field that should be displayed all the time (field is never false) to ignore its required attribute, in particular with one2many's.

After this commit, we differentiate between the rendering function and the validate to write function, making a type of field both requirable and always displayed even if empty

OPW 766002

Closes #1306