[FIX] stock_barcode: picking with products tracked by lot
Before this rev., processing a stock picking with products tracked
by lot with a barcode scanner wasn't working at all. This rev.
aims to re-introduce the behavior as it was in saas-15, i.e.
check in JS if the scanned product is tracked by lot, and if so,
call the route 'get_po_to_split_from_barcode'.

The API of this route has been changed a bit, so that it directly
returns the split lot wizard on the correct PO, instead of
returning the PO id. This saves us an RPC.

Thanks to @abh-odoo for the preliminary work.