[FIX] sale_ebay: don't override partner information
After 15 days, eBay sends 'Invalid Request' as buyer email. So we cannot
make the partner search in Odoo with this value or multiple buyers will
match the same partner.
After this commit, the buyer with 'Invalid Request' as email will be match
with partner in Odoo with the same 'ebay_id'. This means that if the partner
already exists in Odoo it will be duplicated but avoid the lost of information
due to the 'Invalid Request' email.