[FIX] account_reports: followup reports: access rights issue
In multi-company, it may happen that a user has access to a record
of 'account.report.context.followup' but that a message has been
sent on that record by a partner he doesn't has access to. In that
case, before this rev., it crashed when rendering the report, as
it tries to access that partner's name.

Courtesy of @qdp-odoo.
OPW 769993