[REVERT] ee3d8ba3d04f67105ed49295f831253dd797b280
Over-eager removal of xml_declaration in etree.tostring: it must be
removed from serialisations of view architectures as view archs should
be serialised to text (unicode) and lxml refuses to parse text
documents with a document declaration (as the declaration's encoding
can/will conflict with the document's lack of encoding).

However in this case we are not serialising view architectures but
financial documents, so this change was *at best* unnecessary.

Furthermore, according to @moylop260 the Mexican government actually
requires document declarations to be present.