[FIX] web_studio: Custom elem next to custom elems
Previous to this commit, adding a custom element in studio adjacent
to another custom element would put all of the elements in an xpath
with position 'inside' and thus it would break the display of elements.

This was because during the normalization process, the method in
charge of checking for an xpath to anchor to was only looking one
level deep (next node or previous node), so if the adjacent nodes
were studio nodes, it would not be able to anchor to them and would
drop to an else clause which would change the position to 'inside'.

This is fixed by creating a loop that would iterate over the adjacent
nodes (next and previous) and if none were anchorable, it would go
one level deeper (next's next, previous' previous) until it finds one,
if none are found, then we do drop to the else where the 'inside' position
is chosen.