[FIX] l10n_mx_edi: Python 3 does not support the ur prefix
Thankfully for this case the "raw" part is not necessary, just remove it.

Using a "native string" pattern is not an option, it has different
matching semantics than a unicode pattern (and I assume the unicode
pattern was used on purpose). The alternative would be to use a `br`
prefix (supported on both Python 2 and Python 3) and decode() it
immediately before passing it to `re.compile`.