[IMP] marketing_automation: Improve navigation, layout and wording
Purpose
=======

Currently the module is not easy to use.

Specification
=============

Models
~~~~~~

The models have been modified and rename to be more intuitive
functionally and technically. Functionnaly the models are the
following.

1/ The campaign definition:

- The model marketing.campaign defines the principal variables that
  will be used during the campaign.
	- The target model: On which model emails are sent/operations are
	  made (eg: Customers, Leads)
	- Filter: Make operation on a subset of the specified model
	- Unique field: Specify a field that is used to avoid to make
	  operations on 2 records with the same value for this field
	  (eg:  email address)
- The model marketing.activity defines an activity that could be
  executed on the campaign.
	- It can either 1/ send an eamil or 2/ execute a server action
	- Could be performed x (hours, days, weeks, months) after a
	  certain trigger
	- The trigger could be:
		1. The beginning of the campaign
		2-3. the mail is (not) opened
		4-5. the mail is (not) replied
		6-7. the mail is (not) clicked
		8. the mail is bounced
		9. after a given activity
	- The activity could have a validity duration
	- A additional domain could be applied
	- A hierarchy structure could be defined on the activities. We
	  could have the following use case
		Activity A: Sent a 'Welcom' email with a link to a trial
			Activity A1: If not opened after a week, send another
			             email
			Activity A2: If opened after a week, assign a sales
			             person

2/ The campaign execution
- The model marketing.participant defines the record that will follow
  the defined campaign and activities.
	- It linked to the campaign and a given record that match the
	  model defined on the campaign
- The model marketing.trace defines the activities that the particant
  follows, could follow and could have followed.
	- When a participant is generated, all the beginning activities
	  generates related traces for him.
	- When a participant executes an activity, the child activities
	  generate related traces for him.
	- If a participant executes an activity, the opposite activity is
	  canceled for him (eg: mail clicked >< mail not clicked)

Automation
~~~~~~~~~~

The module is automated with 2 crons:

1/ Participans generation:

This cron creates the participant for the running campaign. When a
campaign is launched, all the valid participants are generated with
the beginning traces. When a campaign is already launched since a
certain amount of time, all the new valid participants are generated
too. It means that a short term campaign can be launched during a
long time, every record will follow the same path.

2/ Activity execution / Traces management:

This cron manages the traces execution, and creates the new one when
required.

Activity Modifications
~~~~~~~~~~~~~~~~~~~~~~

When the activities are modified or created on a running campaign,
the user has the possibility to update the participants, i.e. update
the traces scheduled dates, or add them new traces related to the new
activities.

Brand new display
~~~~~~~~~~~~~~~~~

The campaign configuration was before a long and complicated journey.
This commit bring a all new layout to manage the campaign activities
is an ergonomic and beautiful way that allows to track efficiently
the campaign evolution and bottlenecks.