[IMP] web_enterprise: add withelist class to display element in app switcher
This commit introduces a class 'o_in_appswitcher' which will whitelist the
element and let this element displayed in the app switcher.
This commit adds this class to the voip panel as well.

Fix in communiy done with the commit odoo/odoo@43e412ccabef7a04c9513d74da5436a57d7ff2ca.