[IMP] account_reports: add 'count_rows' and 'from_context' in formula evaluation
'count_rows': new operator that allows to count the number of rows returned by the SQL query defined on the same line through 'domain' and 'groupby' fields.
'from_context: new operator that allows to retrieve value of a financial line as explicitly given in the context. Useful in case we want to fill information in a wizard and then call a report with the user's encoded values. See l10n_es_report for a complete example.

[FIX] account_reports: filter_comparison=None wasn't supported (was giving traceback)