[FIX] web_enterprise: restore kanban LESS rules
/!\ Should not be forward-ported in saas-18

Before saas-16 (the new views), the kanban view was in a separated
app (web_kanban). This installed its assets after the web_enterprise
app. In particular, the web_enterprise/kanban_view.less file was
placed before the web_kanban/kanban_view.less. As web_enterprise
should be able to override web_kanban, the solution that seemed to
have been chosen was to increase the specificity of the web_enterprise
LESS rule to make sure it overrode the web_kanban LESS rules (the
correct solution would have been to make web_enterprise depends on
web_kanban).

Since web_kanban is now part of the web app, the order was changed
and some web_enterprise rules were overridding rules it should not.
This commit decreases the web_enterprise rules specificity to solve
the problem (one example is the folded columns' hovering effects which
were not occuring anymore).