[FIX] web_studio, *: appswitcher background & navbar transparency
* web_enterprise

Since the new design, the top menu is transparent when the appswitcher
is opened. Studio allows to change the appswitcher background. Since we
cannot predict the type of image the user will choose as background,
transparent navbar might be an issue in that case. So, if a custom
background is set, disable the navbar transparency.