[IMP] web: improve renderCheckbox to be re-usable in front end
Related to #b94939377a7dee66f7860f4ef9e59be85cccf2ba

Before this commit, the label was into a display block that add a return line
between the cehckbox and the label.

It was not visible in backend, because label by defaut is displayed without
renderCheckbox. (no options.text provided)

Now renderCheckbox render correctly with the same code in frontend and backend