[FIX] website_version: publishing a version was deleting related website.page
Before this commit, when you published a version of a page (which is an
ir.ui.view record) it deleted the ir.ui.view's website.page record.

This bug was introduced with the new model website.page and the first
fix of website_version. Indeed, publishing a version is basically copying
ir.ui.view(s) and deleting the original ones.
But with ir.ui.view record now being linked to website.page records (inherits),
deleting an ir.ui.view will delete it's website.page. This is not what we want
in this case.

Now publishing a version will update the websit.page record before deleting
the ir.ui.view.