[ADD] website_twitter_wall: Add a new module to manage walls for events
Purpose
=======

Add a way to manage a twitter wall when organizing events.

By simply entering your twitter credentials and configuring a wall,
add a simple way to display lazily the tweets on a dedicated page.

The tweets are displayed incrementally when they are posted on twitter
if they match the rules defined on the wall.

The tweets are stored on the database. The wall can thus be shared after
the event to show a story.

Co-authored-by: Martin Trigaux <me@mart-e.be>
Co-authored-by: Ravi Gadhia <rga@odoo.com>
Co-authored-by: Denis Ledoux <dle@odoo.com>