[FIX] web_studio: correct use of self
Before this rev., it was not possible to open a calendar view
with Studio if the view has been previously opened.

This was due to the lazy loading of the view librairies, which
was not done correctly in Studio since rev.
https://github.com/odoo/enterprise/commit/d864717fa29c4a7c39fac32d22a3ee7b5807ab4f

Note that this behaviour can't be tested as the librairies lazy
loading is not applied in the tests (the librairies are directly
loaded in the QUnit suite template)