[FIX] web_studio: edit a field in Search Editor
Before this rev., clicking on a field in the Search Editor was causing a traceback.

As the search view is not really a view like the others, there is no attribute
`state` on it.

This lookup was however not useful as the variable is not used when editing the
search view.