[IMP] mrp_plm: better display images in set cover widget
The o_image class is meant for attachments when e.g. uploading a file
on a message in the chatter. It uses the mime type of the image to
display a certain thumbnail that represents that type. It shouldn't be
used when actually displaying the real image with <img>. Instead use
img-responsive which will make the images scale nicely on displays of
different sizes.