[FIX] marketing_automation: activity display order
In the Campaign view, activities were displayed in the wrong order:
activities that started a few "days" after the beginning of the campaign
were displayed before activities starting "hours" after the beginning of
the campaign (simply because "days" is before "hours" in the
lexicographic order).

We introduce a standardized value for all activities (in hours) that
allow to sort them properly.