[FIX] web_studio: export automations
The rev. https://github.com/odoo/odoo/commit/55549f1e612793769fc7bdf860ac6af06735a755
removed the model `ir.values` and all its references.
The field `menu_ir_values_id` has thus been deleted so there is no need to
mention it in the field not to export in Studio.

Note that this field has been replaced with new ones `binding_model_id` and
`binding_type`. The first one is a Many2one linked to `ir.model` which is an
exported model (`ir.value` was not exported, hence its presence in field not to
export). The second one is a selection so it can also be exported.