[IMP] mrp_plm: compute effectivity based on effectivity_date
When creating an ECO through the calendar view you end up with an ECO
with an effectivity_date and an effectivity of 'asap' which doesn't
make much sense. The calendar view automatically sets effectivity_date
which is ok, but the default of effectivity is 'asap'. Also, when
setting an effectivity_date it always shows up in the kanban view,
even for the 'asap' effectivity.

Since these two fields are so closely related and because an
effectivity of 'asap' with an effectivity_date doesn't make much sense
let's make effectivity a stored computed field that depends on
effectivity_date. Also provide an inverse so people can still set it.

Note that the default effectivity of 'asap' remains the same. An ECO
by default has no effectivity_date so the default effectivity remains
'asap'.

Closes #1418