[FIX] project_forecast: no task break ir rule
In 0c59fbd8 was added a multi company rule, but if there is no task this
will block weird things.

It seems there should be a task but an access rule is not the place to
do something like this, and since this unwanted side effect of the
above-mentioned change I simply fix it.

opw-771765
closes #1457