[FIX] l10n_nl_reports: fix rounding error in XAF report
The XAF report only support 2 decimal places on monetary values.
In our case, some account.move.line could have a debit or credit having a huge number of decimal places.

How is it possible because debit/credit are monetary fields?

Because the float_round method on tools is not perfect. Indeed, some decimal values are not well truncated in the method.
For example, float_round(53.550000000000004, precision_rounding=0.01) results 53.550000000000004 instead of 53.55.