[FIX] web_studio: correctly create a `priority` widget
Since the rev. https://github.com/odoo/enterprise/commit/57a6585a2d31f2530182b0a1c13701607abe3286 we removed
all attributes passed except the name. This was not exactly correct as it's
possible to drag and drop a field with the attribute `widget` already set.

This was thus not possible anymore to create a priority widget directly.