[FIX] helpdesk: Add missing triggers on compute method
As the method '_compute_sla_fail' uses 'sla_id.stage_id.sequence'
and 'stage_id.sequence' their values are set to false and are
recomputed if not set as triggers of the method.

The field deadline is set to False too by accessing to sla_id
(that triggers the method '_compute_sla'), leading to a traceback
by comparing a date with False.