[FIX] sale_subscription: Upsell of a not consumed subscription
If the starting date of the subscription is equal to the next invoice date,
it means that there is nothing to invoice.

time_to_invoice = recurring_next_invoice - date - datetime.timedelta(days=1)

For example if recurring_next_invoice = 02/10/2017, date = 02/10/2017 and timedelta = 1 day,
recurring_last_invoice = 02/10/2016

Then time_to_invoice = -1

So the ratio was equal to -1 / recurring_next_invoice - recurring_last_invoice (equal to 365)

And the discount on the SO was equal to -100,27% which was wrong.