[FIX] account_reports: propagate the active_id in action context (#1474)
When clicking on a account.financial.html.report.line having an action_id, the action is well triggered but we don't have the
active_id of the clicked line. It leads to a traceback in some action's code, e.g. action_journal_items from GST report (l10n_au_reports).

To avoid that, the active_id is now added in the action's context during the execute_action.