[IMP] stock_barcode: is_done field does not exist
Since stock refactoring is_done field does not exist anymore.
In 10.0 it was computed as qty_done > 0. Since the check qty_done
was already present when is_done was used we removed it.

This commit also allow to add the quantity reserved as quantity
done when the user scan a package that is the source and destination
for a stock move line.