[FIX] quality_mrp: alter skip button behaviour in tablet view
Previously, when we were on a step which was generated on-the-fly,
clicking previous then skip did not make us come back to the expected
quality check.

Similarly, clicking 'finish steps' (which calls
action_first_skipped_step) did not work if the first skipped step had
been generated on-the-fly.

This commit fixes both of the aforementioned issues.