[FIX] web_studio: correctly add a `group_by`
It was not possible to add a field in the `group_by` section in the
search view if the tag <group> was not already in the view before
this rev.

When adding a field in the `grouby` section, the attribute `create_group`
is passed and popped in the controller after the group creation.

This attribute thus needs to be picked.