[FIX] web_gantt: duration of bars
dhtmlxgantt's `calculateDuration` apparently always returns an
integral amount (of work hours) which leads to completely broken stuff
when giving it times which are not "on the dot" leading to visual
overlaps (or gaps) between steps which depend on one another e.g. on a
work order with one step from 13:12 to 14:38 and the second one from
14:38 to 16:04, the duration of the first phase would be computed to
2, and thus displayed (on the gantt) from 13:12 to 15:12, leading to
~25% overlap with the following step.

dhtmlx apparently handles non-integral durations just fine, so
handroll the computation in order to fix the visual display.

OPW-771373