[FIX] account_taxcloud: Stop magic guessing the tax rates
Purpose
=======

Currently in Odoo when mapping taxes, we check there is a fiscal position line corresponding the SO or Invoice line according to the zip, the state and the TIC category.

In fact, tax rates depend on several factors:
- The State, zip, street1, street2, ...
- The unit price and quantity
- The TIC category
- The time

Here we are guessing the rate when encoding the line without taking all the factors into account. The fact that the tax rates are depending of the time is also another argument to prove that this way to work is obviously wrong.

We shouldn't make an approximation of the real tax rate to recompute it on the SO confirmation. This is confusing for the user.

Specification
=============

Remove all the code corresponding to the tax rates mapping from the fiscal position.

The fields on the fiscal position could be removed from the model in the master version.