[FIX] (sale_)account_taxcloud: Display 'Sales Tax' in SO/Inv reports
Purpose
=======

Context
-------

The tax rate amount is infered from the taxed amount returned by taxcloud. UNfortunately there is a loss of information due to the following reasons:

- Taxcloud returns the rounded tax amount (2 decimals)
- A sales tax rate could have a decimal precision of 3 (See New Mexico with tax rates of 6.873%)
- Odoo uses the model account.tax and thus the tax rate is mandatory to compute the tax amount.

This loss of information could lead to a tax rate that is different from the tax rate you could find on the taxcloud interface (https://taxcloud.net/account/#Find-a-Rate)

Example
-------

NB: Creadentials for the example to (https://github.com/bazemoreb).

Create a partner with the address of 1521 S Burr St, Mitchell, SD 57301 (tax rate should be 6.5% - confirmed on taxcloud's website). Now create a product with price of $25 and a product with price of $19.95. Add those to either a quotation or an invoice (it seems to follow same steps and same result). Now confirm quote or validate invoice.

What is the current behavior that you observe?
After a confirm/validate, this tax rate is set to 6.52%. I've figured out why I think.

My example uses the two item lines I referenced above:
Voip seats - 25.00
Plantronics adapter – 19.95

If you multiply by 6.5% you get:
25*0.065 = 1.625
19.95*0.065 = 1.29675
If you then round that to two digits you get 1.63 and 1.30 respectively.

Then you divide those BACK in:
1.63/25 = 6.52%
1.3/19.95 = 6.516% (rounds to 6.52%)

That way, the taxed amount is
25*0.0652 = 1.63
19.95*0.00652 = 1.30074
If you then round that to two digits you get 1.63 and 1.30 respectively.

Issue
-----

The problem here is that even if the taxed amount is correct, the user may think that the tax rate is incorrect as it is not equal to 6.5%.

Specification
=============

On the sales order report and on the invoice report, we should avoid the user to get confused by displaying 'Sales Tax' instead of 'Tax 6.52%'.

We can't do more than that as the tax name is subject to a unique constraint.